change mobile 'my location' to zoom level 15, rather than 13
ideally, the zoom level should be preserved when locating, perhaps with minimum and maximum zoom levels. this way it'd work well for both low portal density regions and the higher portal density in cities
(i.e. at least zoom level 13, no more than zoom level 17). leaflet doesn't allow for this as it stands, so tis is a quick-n-dirty improvement