Create PortalPath.js
Creates a file that contains a list of portals in the order you clicked on them. The file contains a sequence number, the portal name, a URL that should start the navigation app on a phone. So click on the portals in the order you want to visit them. Save the file. Copy the contents of the file into an email message and send to your phone. Then you can click on links and run navigation in the background while you ingress.