Show link range details for partially deployed portals.
Calculates the link range for portals with <8 resonators as well, but uses a dashed circle and a striked-through range value for those.

The tooltip over the range value in the details pane shows base range, link amp boost and total range.

The range for portals with missing resonators can be important because links with shorter lengths might originate here. Those will collapse when the total link range drops below their length, but not as soon as a resonator is destroyed.