Fix the calculation for Link Amp range boost to be generic, so VRLA boosts are correctly calculated.
For RLA, the first multiplier is 2x. Subsequent RLAs add 0.5, 0.25, 0.25 to the multiplier. But expressed as a fraction of the initial multiplier, this is 0.25, 0.125, 0.125. Scale array has been updated to reflect this. This means the "-1" in the calculation is no longer required, and that boost is correctly calculated when baseMultiplier is anything other than 2 (eg VRLA, when it is 7).
If there aren't any Link Amps, the multiplier is 1x.