Update draw-tools.user.js
Currently if you draw a small polygon, then a large one, if you try to delete the small polygon the large polygon is deleted because it covers the small polygon. This mod sorts the drawn items by area so the smallest polygons and lines are above larger polygons so they are properly clickable.