Variation of reso energy % for days remaining
A simple adaption of reso-energy-pct-in-portal-details.user.js where instead of displaying the remaining percentage, it calculates the decay rate (15% of max per day) and displays the (pessimistic) number of days remaining based on the remaining energy. 

Issues: If both days and % scripts are enabled, they will both display and look garbled. There are plenty of ways to work around this but as this is my first contribution to IITC and UI dev is not generally my area of expertise, I'm happy to take suggestions, or for someone else to roll it in to the pct script.