Move and zoom to portal double clicking its title.
Move and zoom to portal double clicking its title in portal detail view. 
Portal click in chat doesn't center a portal and sometimes it apeears under the chat (sometimes portal selection doesn't work at all for me). Another situation is when you moved the map away from the selected portal and want to see it again. Anyway if you see portal info you should have the ability to move to it simply.

Also full name of a portal is in hint now. It is useful in case it is too long and was truncated.