Changed the colour of ENL ant-trails, made dashes asymmetric.
Changed the colour of the ENL ant-trail in order to make it easier to distinguish between ENL and RES agents when trying to coordinate events.
Made the dash-pattern slightly asymetric so that it would be possible-ish to determine which way players move (and arrow would be better but that would be a bigger change).