Trims second and minutes
This will trim seconds and minutes if they both end up being :00.  If
local time isn't a whole hour step from UMT, the minutes won't be :00
and instead will be :30 or whatever the step is while still trimming the
seconds.  If seconds are somehow not :00, they'll remain.  The lookahead
is to keep am/pm from being stripped.